requirements.txt
requests==2.31.0
python-dotenv==1.0.0
apify-client==1.7.1
prisma==0.11.0
